clean up predict method
