+++ src/tasks/predict_modality/methods/simple_mlp/predict/config.vsh.yaml
@@ -1,29 +1,10 @@
-__merge__: ../../api/comp_method.yaml
+__merge__: ../../../api/comp_method_predict.yaml
 functionality:
-  name: simplemlp
-  info:
-    label: Simple MLP
-    summary: Ensemble of MLPs trained on different sites
-    description: Ensemble of MLPs trained on different sites
-    documentation_url: https://github.com/openproblems-bio/neurips2021_multimodal_topmethods/tree/main/src/predict_modality/methods/novel#readme
-    repository_url: https://github.com/openproblems-bio/neurips2021_multimodal_topmethods/tree/main/src/predict_modality/methods/novel
-    reference: pmlr-v176-lance22a     
-  arguments:
-    - name: "--input_model"
-      type: "file"
-      example: "pretrain_model"
-      description: Path to the directory containing a pretrained model.
-      required: true
-    - name: "--output"
-      type: "file"
-      direction: "output"
-      example: "output.h5ad"
-      description: Dataset with predicted values for modality2.
-      required: true
+  name: simplemlp_predict
   resources:
     - type: python_script
       path: script.py
-    - path: helper_functions/
+    - path: ../resources/
 platforms:
   - type: docker
     image: pytorch/pytorch:1.9.0-cuda11.1-cudnn8-runtime
@@ -36,4 +17,4 @@ platforms:
           - pytorch-lightning
   - type: nextflow
     directives:
-      label: [ highmem, hightime, highcpu, gpu]
+      label: [highmem, hightime, highcpu, gpu]
